Alteração para código ficar de acordo com phpstan

+++ app/Operations/SquareRoot.php
@@ -4,9 +4,9 @@ namespace Calculator\Operations;
 
 class SquareRoot extends OperationAbstract {
 
-    public function calc():float
+    public function calc(): float
     {
-        return (float) sqrt($this->number1);
+        return sqrt($this->number1);
     }
 
     public function __construct(float $number)
